show all subtests in qualification overview page, even when subtests are in different folders. add option to also show outdated results (instead of overwriting by the latest). don't show duplicate comments on overview pages. sort qualification overview page results descending by date

+++ Analyse/TestResultClasses/CMSPixel/QualificationGroup/Fulltest/Fulltest.py
@@ -903,9 +903,10 @@ class TestResult(GeneralTestResult):
             if self.verbose:
                 print "INSERT ROW:", Row
             with self.TestResultEnvironmentObject.LocalDBConnection:
-                self.TestResultEnvironmentObject.LocalDBConnectionCursor.execute(
-                    'DELETE FROM ModuleTestResults WHERE ModuleID = :ModuleID AND TestType=:TestType AND QualificationType=:QualificationType AND TestDate <= :TestDate',
-                    Row)
+                if self.TestResultEnvironmentObject.Configuration['ShowOnlyLatestTestResults']:
+                    self.TestResultEnvironmentObject.LocalDBConnectionCursor.execute(
+                        'DELETE FROM ModuleTestResults WHERE ModuleID = :ModuleID AND TestType=:TestType AND QualificationType=:QualificationType AND TestDate <= :TestDate',
+                        Row)
                 self.TestResultEnvironmentObject.LocalDBConnectionCursor.execute(
                     '''INSERT INTO ModuleTestResults
                     (
